feat(frontend/app/page.tsx): Created the Bids page that shows all the products a user placed a bid on
Currently, i'm only fetching the products from one user and it's based on an integer i pass inside the route. The backend doesn't seem to be getting the token from the user, and for some reason this bug only happens on the GET request and not the POST/DELETE requests.