feat(frontend/): Redone the entire Auth setup
Simplified the Authentication system by fully relying on cookies.

Changed majority of the components to be server components instead of client components. Components such as Header are still not working as expected but will find a work around.

Axios also needs the Authentication headers on each request, especially for the protected routes as I still haven't managed to make it work.